Provide a method for configuring SSH connections
The manager didn't allow one to specify options for the SSH connection
besides the port to connect to on the remote host.  To address this
issue I added a mechanism which will pass the SSH options specified in
the configuration files to ssh/scp commands used internally by the
manager.